+++ src/o365/spo/common/base-permissions.ts
